salt-master-management/requirements.txt
# Note this is part of py2 --> py3 upgrade (3.7 to be exact)
# previous versions in py2
#
# gitdb==0.6.4
# GitPython==1.0.1
# smmap==0.9.0
#
# The following are latest as of 6/2020:
# pip3 install gitpython
gitdb==4.0.5
GitPython==3.1.3
smmap==3.0.4
